# docs/api.yml
openapi: 3.0.0
info:
  version: "latest"
  title: PROG3-TD2
paths:
...
  /players:
    get:
      tags:
        - Players
      summary: get all players
      operationId: getPlayers
      responses:
        '200':
          description: List of players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Player'
        '500':
          description: Internal server error
    post:
      tags:
        - Players
      summary: create new players
      operationId: createPlayers
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CreatePlayer'
      responses:
        '200':
          description: List of created players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Player'
        '500':
          description: Internal server error
    put:
      tags:
        - Players
      summary: update some player
      operationId: updatePlayer
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/UpdatePlayers'
      responses:
        '200':
          description: List of updated authors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Player'
        '500':
          description: Internal server error
components:
  schemas:
...
    Player:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
    UpdatePlayers:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        number:
          type: number
        team:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
    CreatePlayer:
      type: object
      properties:
        name:
          type: string
        number:
          type: number
        team:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string